Quiet g++ compiler warnings Work around g++ compiler bug instantiating vector<NCVariable<Vector> > Added computeVariableExtents to (eventually) simplify data warehouses

+++ Uintah/Grid/CCVariable.h
@@ -327,7 +327,7 @@ class CCVariable : public Array3<T>, public CCVariableBase {
 	    IntVector h(getHighIndex());
 	    for(int z=l.z();z<h.z();z++){
 	       for(int y=l.y();y<h.y();y++){
-		  size_t size = sizeof(T)*(h.x()-l.x());
+		  ssize_t size = (ssize_t)(sizeof(T)*(h.x()-l.x()));
 		  ssize_t s=write(oc.fd, &(*this)[IntVector(l.x(),y,z)], size);
 		  if(size != s)
 		     throw ErrnoException("CCVariable::emit (write call)", errno);
@@ -356,7 +356,7 @@ class CCVariable : public Array3<T>, public CCVariableBase {
 	    IntVector h(getHighIndex());
 	    for(int z=l.z();z<h.z();z++){
 	       for(int y=l.y();y<h.y();y++){
-		  size_t size = sizeof(T)*(h.x()-l.x());
+		  ssize_t size = (ssize_t)(sizeof(T)*(h.x()-l.x()));
 		  ssize_t s=::read(oc.fd, &(*this)[IntVector(l.x(),y,z)], size);
 		  if(size != s)
 		     throw ErrnoException("CCVariable::emit (write call)", errno);
@@ -381,6 +381,11 @@ class CCVariable : public Array3<T>, public CCVariableBase {
 
 //
 // $Log$
+// Revision 1.24  2000/09/25 20:37:42  sparker
+// Quiet g++ compiler warnings
+// Work around g++ compiler bug instantiating vector<NCVariable<Vector> >
+// Added computeVariableExtents to (eventually) simplify data warehouses
+//
 // Revision 1.23  2000/09/25 18:12:19  sparker
 // do not use covariant return types due to problems with g++
 // other linux/g++ fixes
