+++ Uintah/Grid/SFCXVariable.h
@@ -310,7 +310,7 @@ class SFCXVariable : public Array3<T>, public SFCXVariableBase {
 	    IntVector h(getHighIndex());
 	    for(int z=l.z();z<h.z();z++){
 	       for(int y=l.y();y<h.y();y++){
-		  size_t size = sizeof(T)*(h.x()-l.x());
+		  ssize_t size = (ssize_t)(sizeof(T)*(h.x()-l.x()));
 		  ssize_t s=write(oc.fd, &(*this)[IntVector(l.x(),y,z)], size);
 		  if(size != s)
 		     throw ErrnoException("SFCXVariable::emit (write call)", errno);
@@ -340,7 +340,7 @@ class SFCXVariable : public Array3<T>, public SFCXVariableBase {
 	    IntVector h(getHighIndex());
 	    for(int z=l.z();z<h.z();z++){
 	       for(int y=l.y();y<h.y();y++){
-		  size_t size = sizeof(T)*(h.x()-l.x());
+		  ssize_t size = (ssize_t)(sizeof(T)*(h.x()-l.x()));
 		  ssize_t s=::read(oc.fd, &(*this)[IntVector(l.x(),y,z)], size);
 		  if(size != s)
 		     throw ErrnoException("SFCXVariable::emit (write call)", errno);
@@ -371,6 +371,11 @@ class SFCXVariable : public Array3<T>, public SFCXVariableBase {
 
 //
 // $Log$
+// Revision 1.7  2000/09/25 20:37:43  sparker
+// Quiet g++ compiler warnings
+// Work around g++ compiler bug instantiating vector<NCVariable<Vector> >
+// Added computeVariableExtents to (eventually) simplify data warehouses
+//
 // Revision 1.6  2000/09/25 18:12:20  sparker
 // do not use covariant return types due to problems with g++
 // other linux/g++ fixes
