+++ Uintah/Grid/SFCZVariable.h
@@ -311,7 +311,7 @@ class SFCZVariable : public Array3<T>, public SFCZVariableBase {
 	    IntVector h(getHighIndex());
 	    for(int z=l.z();z<h.z();z++){
 	       for(int y=l.y();y<h.y();y++){
-		  size_t size = sizeof(T)*(h.x()-l.x());
+		  ssize_t size = (ssize_t)(sizeof(T)*(h.x()-l.x()));
 		  ssize_t s=write(oc.fd, &(*this)[IntVector(l.x(),y,z)], size);
 		  if(size != s)
 		     throw ErrnoException("SFCZVariable::emit (write call)", errno);
@@ -341,7 +341,7 @@ class SFCZVariable : public Array3<T>, public SFCZVariableBase {
 	    IntVector h(getHighIndex());
 	    for(int z=l.z();z<h.z();z++){
 	       for(int y=l.y();y<h.y();y++){
-		  size_t size = sizeof(T)*(h.x()-l.x());
+		  ssize_t size = (ssize_t)(sizeof(T)*(h.x()-l.x()));
 		  ssize_t s=::read(oc.fd, &(*this)[IntVector(l.x(),y,z)], size);
 		  if(size != s)
 		     throw ErrnoException("SFCZVariable::emit (write call)", errno);
@@ -373,6 +373,11 @@ class SFCZVariable : public Array3<T>, public SFCZVariableBase {
 
 //
 // $Log$
+// Revision 1.7  2000/09/25 20:37:43  sparker
+// Quiet g++ compiler warnings
+// Work around g++ compiler bug instantiating vector<NCVariable<Vector> >
+// Added computeVariableExtents to (eventually) simplify data warehouses
+//
 // Revision 1.6  2000/09/25 18:12:20  sparker
 // do not use covariant return types due to problems with g++
 // other linux/g++ fixes
